Fix duration parsing and individual test failure reporting
- Update Go structs to use float64 for duration fields instead of int to support floating-point durations from test runners
- Fix console output formatting to use %.0f instead of %d for duration display
- Update Vitest adapter to populate FailedTests array with actual failed test names using recursive task traversal
- Remove hardcoded "String operations" test output in orchestrator
- Add integration test to verify Vitest adapter reports individual failed test names
- Update all test files to use float64 duration values

+++ internal/ipc/events.go
@@ -70,7 +70,7 @@ type TestCaseEvent struct {
 		TestName  string     `json:"testName"`
 		SuiteName string     `json:"suiteName,omitempty"`
 		Status    TestStatus `json:"status"`
-		Duration  int        `json:"duration,omitempty"`
+		Duration  float64    `json:"duration,omitempty"`
 		Error     string     `json:"error,omitempty"`
 	} `json:"payload"`
 }
@@ -85,7 +85,7 @@ type TestFileResultEvent struct {
 		Status      TestStatus `json:"status"`
 		FailedTests []struct {
 			Name     string `json:"name"`
-			Duration int    `json:"duration,omitempty"`
+			Duration float64 `json:"duration,omitempty"`
 		} `json:"failedTests,omitempty"`
 	} `json:"payload"`
 }
@@ -105,7 +105,7 @@ type TestCase struct {
 	Name     string     `json:"name"`
 	Suite    string     `json:"suite,omitempty"`
 	Status   TestStatus `json:"status"`
-	Duration int        `json:"duration,omitempty"`
+	Duration float64    `json:"duration,omitempty"`
 	Error    string     `json:"error,omitempty"`
 }
